Feat: Enhance UI with layout and improved dashboard
This commit introduces a new `Layout.tsx` component to provide a consistent navigation experience across all pages.

The changes include:
- A new `Layout.tsx` component that renders the `Navbar`.
- `App.tsx` has been updated to use this layout, so all pages now have a navigation bar.
- The `Dashboard.tsx` page has been enhanced with a more structured layout, including a header, a button, and a table with placeholder data.

These changes create a more complete and user-friendly MVP UI.
