--- tenant-management-app/frontend/public
@@ -1,1 +1,0 @@
-<!-- This file is intentionally left blank. -->
