--- tenant-management-app/frontend/src/pages/Dashboard.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Dashboard = () => {
-  return (
-    <div>
-      <h1 className="text-2xl font-bold">Dashboard</h1>
-      <p>Welcome to the tenant management dashboard.</p>
-    </div>
-  );
-};
-
-export default Dashboard;
