+++ tenant-management-app/package.json
@@ -18,6 +18,8 @@
   "devDependencies": {
     "vite": "^3.2.11",
     "typescript": "^4.0.0",
-    "@vitejs/plugin-react": "^2.2.0"
+    "@vitejs/plugin-react": "^2.2.0",
+    "postcss": "^8.4.38",
+    "autoprefixer": "^10.4.19"
   }
 }
