+++ tenant-management-app/src/components/Layout.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import Navbar from './Navbar';
+
+type LayoutProps = {
+  children: React.ReactNode;
+};
+
+const Layout: React.FC<LayoutProps> = ({ children }) => {
+  return (
+    <div className="min-h-screen bg-gray-100">
+      <Navbar />
+      <main className="container mx-auto p-4">{children}</main>
+    </div>
+  );
+};
+
+export default Layout;
