+++ tenant-management-app/src/components/Navbar.tsx
