+++ tenant-management-app/src/pages/Properties.tsx
