+++ tenant-management-app/src/styles/index.css
