+++ tenant-management-app/src/types/index.ts
