+++ tenant-management-app/src/utils/supabaseClient.ts
