+++ tenant-management-app/tsconfig.json
