+++ tenant-management-app/vite.config.ts
